requirements.txt
flask~=3.1.0
psycopg2
authlib~=1.5.2
requests
gunicorn

python-dotenv~=1.1.0
psycopg2-binary~=2.9.9